requirements.txt
browser-use
playwright 
python-dotenv
langchain-google-genai